python_converter/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
chardet==5.2.0
click==8.1.7
colorama==0.4.6
et_xmlfile==2.0.0
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
idna==3.10
numpy==2.1.3
openpyxl==3.1.5
pandas==2.2.3
pyarrow==18.1.0
pydantic==2.10.3
pydantic_core==2.27.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.19
pytz==2024.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.32.1
PyYAML==6.0.2
